hostname and private-address as optional parameters to configure_website()
Some charms are proxies for other charms. Examples of this are charms that
deploy docker containers on an external docker host, or charms that manage an
externally hosted service.

The machine where the Juju agent of a proxy charm runs might not be the machine
where the actual application is running. For this reason, proxy charms might
want to supply a custom hostname and private address. This patch makes that
possible.